refactor(agentPresenter): move permission handler
