+++ src/main/presenter/agentPresenter/permission/index.ts
@@ -1,0 +1,1 @@
+export { PermissionHandler } from './permissionHandler'
