+++ src/main/presenter/sessionPresenter/index.ts
@@ -28,7 +28,7 @@ import { ToolCallHandler } from '../agentPresenter/loop/toolCallHandler'
 import { LLMEventHandler } from '../agentPresenter/streaming/llmEventHandler'
 import { SearchHandler } from '../searchPresenter/handlers/searchHandler'
 import { StreamGenerationHandler } from '../agentPresenter/streaming/streamGenerationHandler'
-import { PermissionHandler } from './permission/permissionHandler'
+import { PermissionHandler } from '../agentPresenter/permission/permissionHandler'
 import { UtilityHandler } from './utility/utilityHandler'
 import { CommandPermissionService } from '../permission/commandPermissionService'
 import type { ThreadHandlerContext } from '../searchPresenter/handlers/baseHandler'
